code for day17 part1

+++ utilities.h
@@ -14,7 +14,7 @@ typedef uint64_t    u64;
 #define set_bit(value, n) ((value) |= ((u64)1<<(n)))
 #define clear_bit(value, n) ((value) &= ~((u64)1<<(n)))
 #define flip_bit(value, n) ((value) ^= (1<<(n)))
-#define check_bit(value, n) ((value) & (1<<(n)))
+#define check_bit(value, n) ((value) & ((u64)1<<(n)))
 
 #define set_bitmask(value, mask) ((value) |= (mask))
 #define clear_bitmask(value, mask) ((value) &= (~(mask)))
